requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.1.0
autobahn==24.4.2
Automat==25.4.16
autopep8==2.3.1
billiard==4.2.1
boto3==1.37.11
botocore==1.37.11
celery==5.5.3
certifi==2024.8.30
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
crispy-bootstrap5==2024.10
cryptography==44.0.0
daphne==4.2.1
decorator==5.2.1
defusedxml==0.7.1
diff-match-patch==20241021
dj-database-url==2.3.0
Django==5.0.9
django-anymail==12.0
django-appconf==1.1.0
django-ckeditor-5==0.2.17
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-cryptography-5==2.0.3
django-environ==0.11.2
django-filter==24.3
django-import-export==4.3.7
django-jazzmin==3.0.1
django-mathfilters==1.0.0
django-recaptcha==4.0.0
django-redis==6.0.0
django-rest-passwordreset==1.5.0
django-silk==5.3.2
django-storages==1.14.5
django-taggit==6.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.8
environs==14.1.1
gprof2dot==2024.6.6
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
incremental==24.7.2
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.4
Markdown==3.7
marshmallow==3.26.1
msgpack==1.1.1
ndg-httpsclient==0.5.1
numpy==2.2.3
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
proglog==0.1.10
prompt_toolkit==3.0.51
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pusher==3.3.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.12.1
pycparser==2.22
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
redis-simple-cache==0.0.8
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
s3transfer==0.11.4
service-identity==24.2.0
setuptools==80.9.0
shortuuid==1.0.13
six==1.17.0
slippers==0.6.2
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.2
stripe==12.1.0
tablib==3.8.0
tqdm==4.67.1
Twisted==25.5.0
txaio==25.6.1
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
zope.interface==7.2
